[frontend] refactor: separate JS script
